Heavily cleaned up code and added type definitions
